flask/requirements.txt
uvicorn[standard]==0.15.0
gunicorn==20.1.0
Flask==2.2.2
python-dotenv==0.21.0